created toggle component + style refactor
