+++ src/assets/ChangeTheme/moonIcon.png
