+++ src/components/localeSwitcher/i18nConfiguration.ts
